Avoid use of `usize` for index

+++ capi/compiler/src/syntax/clusters.rs
@@ -30,4 +30,4 @@ pub struct Cluster {
 ///
 /// Assumes named functions are ordered as they appear in the source code.
 #[derive(Clone, Copy, Eq, Ord, PartialEq, PartialOrd)]
-pub struct NamedFunctionIndex(pub usize);
+pub struct NamedFunctionIndex(pub u32);
